+++ sdk/recoveryservicesbackup/arm-recoveryservicesbackup/src/models/index.ts
@@ -123,6 +123,7 @@ export type AzureVmWorkloadProtectedItemUnion =
   | AzureVmWorkloadProtectedItem
   | AzureVmWorkloadSAPAseDatabaseProtectedItem
   | AzureVmWorkloadSAPHanaDatabaseProtectedItem
+  | AzureVmWorkloadSAPHanaDBInstanceProtectedItem
   | AzureVmWorkloadSQLDatabaseProtectedItem;
 export type AzureWorkloadRecoveryPointUnion =
   | AzureWorkloadRecoveryPoint
@@ -166,6 +167,8 @@ export type AzureVmWorkloadProtectableItemUnion =
   | AzureVmWorkloadSAPAseSystemProtectableItem
   | AzureVmWorkloadSAPHanaDatabaseProtectableItem
   | AzureVmWorkloadSAPHanaSystemProtectableItem
+  | AzureVmWorkloadSAPHanaDBInstance
+  | AzureVmWorkloadSAPHanaHSR
   | AzureVmWorkloadSQLAvailabilityGroupProtectableItem
   | AzureVmWorkloadSQLDatabaseProtectableItem
   | AzureVmWorkloadSQLInstanceProtectableItem;
@@ -652,14 +655,21 @@ export interface ProtectedItem {
     | "AzureVmWorkloadProtectedItem"
     | "AzureVmWorkloadSAPAseDatabase"
     | "AzureVmWorkloadSAPHanaDatabase"
+    | "AzureVmWorkloadSAPHanaDBInstance"
     | "AzureVmWorkloadSQLDatabase"
     | "DPMProtectedItem"
     | "GenericProtectedItem"
     | "MabFileFolderProtectedItem";
-  /** Type of backup management for the backed up item. */
-  backupManagementType?: BackupManagementType;
-  /** Type of workload this item represents. */
-  workloadType?: DataSourceType;
+  /**
+   * Type of backup management for the backed up item.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly backupManagementType?: BackupManagementType;
+  /**
+   * Type of workload this item represents.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly workloadType?: DataSourceType;
   /** Unique name of container */
   containerName?: string;
   /** ARM ID of the resource to be backed up. */
@@ -961,6 +971,8 @@ export interface WorkloadProtectableItem {
     | "SAPAseSystem"
     | "SAPHanaDatabase"
     | "SAPHanaSystem"
+    | "SAPHanaDBInstance"
+    | "SAPHanaHSR"
     | "SQLAvailabilityGroupContainer"
     | "SQLDataBase"
     | "SQLInstance";
@@ -1122,6 +1134,12 @@ export interface SubProtectionPolicy {
   schedulePolicy?: SchedulePolicyUnion;
   /** Retention policy with the details on backup copy retention ranges. */
   retentionPolicy?: RetentionPolicyUnion;
+  /**
+   * Tiering policy to automatically move RPs to another tier.
+   * Key is Target Tier, defined in RecoveryPointTierType enum.
+   * Tiering policy specifies the criteria to move RP to the target tier.
+   */
+  tieringPolicy?: { [propertyName: string]: TieringPolicy };
 }
 
 /** Base class for backup schedule. */
@@ -1140,10 +1158,40 @@ export interface RetentionPolicy {
   retentionPolicyType: "LongTermRetentionPolicy" | "SimpleRetentionPolicy";
 }
 
+/**
+ * Tiering Policy for a target tier.
+ * If the policy is not specified for a given target tier, service retains the existing configured tiering policy for that tier
+ */
+export interface TieringPolicy {
+  /**
+   * Tiering Mode to control automatic tiering of recovery points. Supported values are:
+   * 1. TierRecommended: Tier all recovery points recommended to be tiered
+   * 2. TierAfter: Tier all recovery points after a fixed period, as specified in duration + durationType below.
+   * 3. DoNotTier: Do not tier any recovery points
+   */
+  tieringMode?: TieringMode;
+  /**
+   * Number of days/weeks/months/years to retain backups in current tier before tiering.
+   * Used only if TieringMode is set to TierAfter
+   */
+  duration?: number;
+  /**
+   * Retention duration type: days/weeks/months/years
+   * Used only if TieringMode is set to TierAfter
+   */
+  durationType?: RetentionDurationType;
+}
+
 /** Additional information on Azure IaaS VM specific backup item. */
 export interface AzureIaaSVMProtectedItemExtendedInfo {
-  /** The oldest backup copy available for this backup item. */
+  /** The oldest backup copy available for this backup item across all tiers. */
   oldestRecoveryPoint?: Date;
+  /** The oldest backup copy available for this backup item in vault tier */
+  oldestRecoveryPointInVault?: Date;
+  /** The oldest backup copy available for this backup item in archive tier */
+  oldestRecoveryPointInArchive?: Date;
+  /** The latest backup copy available for this backup item in archive tier */
+  newestRecoveryPointInArchive?: Date;
   /** Number of backup copies available for this backup item. */
   recoveryPointCount?: number;
   /** Specifies if backup policy associated with the backup item is inconsistent. */
@@ -1273,8 +1321,14 @@ export interface AzureStorageJobTaskDetails {
 
 /** Additional information on Azure Workload for SQL specific backup item. */
 export interface AzureVmWorkloadProtectedItemExtendedInfo {
-  /** The oldest backup copy available for this backup item. */
+  /** The oldest backup copy available for this backup item across all tiers. */
   oldestRecoveryPoint?: Date;
+  /** The oldest backup copy available for this backup item in vault tier */
+  oldestRecoveryPointInVault?: Date;
+  /** The oldest backup copy available for this backup item in archive tier */
+  oldestRecoveryPointInArchive?: Date;
+  /** The latest backup copy available for this backup item in archive tier */
+  newestRecoveryPointInArchive?: Date;
   /** Number of backup copies available for this backup item. */
   recoveryPointCount?: number;
   /** Indicates consistency of policy object and policy applied to this backup item. */
@@ -2313,26 +2367,41 @@ export type AzureIaaSVMProtectedItem = ProtectedItem & {
     | "AzureIaaSVMProtectedItem"
     | "Microsoft.ClassicCompute/virtualMachines"
     | "Microsoft.Compute/virtualMachines";
-  /** Friendly name of the VM represented by this backup item. */
-  friendlyName?: string;
-  /** Fully qualified ARM ID of the virtual machine represented by this item. */
-  virtualMachineId?: string;
+  /**
+   * Friendly name of the VM represented by this backup item.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly friendlyName?: string;
+  /**
+   * Fully qualified ARM ID of the virtual machine represented by this item.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly virtualMachineId?: string;
   /** Backup status of this backup item. */
   protectionStatus?: string;
   /** Backup state of this backup item. */
   protectionState?: ProtectionState;
-  /** Health status of protected item. */
-  healthStatus?: HealthStatus;
+  /**
+   * Health status of protected item.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly healthStatus?: HealthStatus;
   /** Health details on this backup item. */
   healthDetails?: AzureIaaSVMHealthDetails[];
   /** Health details of different KPIs */
   kpisHealths?: { [propertyName: string]: KPIResourceHealthDetails };
   /** Last backup operation status. */
   lastBackupStatus?: string;
-  /** Timestamp of the last backup operation on this backup item. */
-  lastBackupTime?: Date;
-  /** Data ID of the protected item. */
-  protectedItemDataId?: string;
+  /**
+   * Timestamp of the last backup operation on this backup item.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly lastBackupTime?: Date;
+  /**
+   * Data ID of the protected item.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly protectedItemDataId?: string;
   /** Additional information for this backup item. */
   extendedInfo?: AzureIaaSVMProtectedItemExtendedInfo;
   /** Extended Properties for Azure IaasVM Backup. */
@@ -2358,17 +2427,24 @@ export type AzureVmWorkloadProtectedItem = ProtectedItem & {
     | "AzureVmWorkloadProtectedItem"
     | "AzureVmWorkloadSAPAseDatabase"
     | "AzureVmWorkloadSAPHanaDatabase"
+    | "AzureVmWorkloadSAPHanaDBInstance"
     | "AzureVmWorkloadSQLDatabase";
-  /** Friendly name of the DB represented by this backup item. */
-  friendlyName?: string;
+  /**
+   * Friendly name of the DB represented by this backup item.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly friendlyName?: string;
   /** Host/Cluster Name for instance or AG */
   serverName?: string;
   /** Parent name of the DB such as Instance or Availability Group. */
   parentName?: string;
   /** Parent type of protected item, example: for a DB, standalone server or distributed */
   parentType?: string;
-  /** Backup status of this backup item. */
-  protectionStatus?: string;
+  /**
+   * Backup status of this backup item.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly protectionStatus?: string;
   /** Backup state of this backup item. */
   protectionState?: ProtectionState;
   /** Last backup operation status. Possible values: Healthy, Unhealthy. */
@@ -2678,6 +2754,12 @@ export type AzureIaaSVMProtectionPolicy = ProtectionPolicy & {
   schedulePolicy?: SchedulePolicyUnion;
   /** Retention policy with the details on backup copy retention ranges. */
   retentionPolicy?: RetentionPolicyUnion;
+  /**
+   * Tiering policy to automatically move RPs to another tier
+   * Key is Target Tier, defined in RecoveryPointTierType enum.
+   * Tiering policy specifies the criteria to move RP to the target tier.
+   */
+  tieringPolicy?: { [propertyName: string]: TieringPolicy };
   /** Instant RP retention policy range in days */
   instantRpRetentionRangeInDays?: number;
   /** TimeZone optional input as string. For example: TimeZone = "Pacific Standard Time". */
@@ -3125,6 +3207,8 @@ export type AzureVmWorkloadProtectableItem = WorkloadProtectableItem & {
     | "SAPAseSystem"
     | "SAPHanaDatabase"
     | "SAPHanaSystem"
+    | "SAPHanaDBInstance"
+    | "SAPHanaHSR"
     | "SQLAvailabilityGroupContainer"
     | "SQLDataBase"
     | "SQLInstance";
@@ -3150,7 +3234,7 @@ export type AzureVmWorkloadProtectableItem = WorkloadProtectableItem & {
 };
 
 /** Azure IaaS VM workload-specific Health Details. */
-export type AzureIaaSVMHealthDetails = ResourceHealthDetails & {};
+export type AzureIaaSVMHealthDetails = ResourceHealthDetails;
 
 /** Log policy schedule. */
 export type LogSchedulePolicy = SchedulePolicy & {
@@ -3219,7 +3303,7 @@ export type SimpleRetentionPolicy = RetentionPolicy & {
 };
 
 /** RecoveryPoint Tier Information V2 */
-export type RecoveryPointTierInformationV2 = RecoveryPointTierInformation & {};
+export type RecoveryPointTierInformationV2 = RecoveryPointTierInformation;
 
 /** Azure Recovery Services Vault specific protection intent item. */
 export type AzureWorkloadAutoProtectionIntent = AzureRecoveryServiceVaultProtectionIntent & {
@@ -3251,6 +3335,12 @@ export type AzureVmWorkloadSAPAseDatabaseProtectedItem = AzureVmWorkloadProtecte
 export type AzureVmWorkloadSAPHanaDatabaseProtectedItem = AzureVmWorkloadProtectedItem & {
   /** Polymorphic discriminator, which specifies the different types this object can be */
   protectedItemType: "AzureVmWorkloadSAPHanaDatabase";
+};
+
+/** Azure VM workload-specific protected item representing SAP HANA DBInstance. */
+export type AzureVmWorkloadSAPHanaDBInstanceProtectedItem = AzureVmWorkloadProtectedItem & {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  protectedItemType: "AzureVmWorkloadSAPHanaDBInstance";
 };
 
 /** Azure VM workload-specific protected item representing SQL Database. */
@@ -3433,6 +3523,18 @@ export type AzureVmWorkloadSAPHanaDatabaseProtectableItem = AzureVmWorkloadProte
 export type AzureVmWorkloadSAPHanaSystemProtectableItem = AzureVmWorkloadProtectableItem & {
   /** Polymorphic discriminator, which specifies the different types this object can be */
   protectableItemType: "SAPHanaSystem";
+};
+
+/** Azure VM workload-specific protectable item representing SAP HANA Dbinstance. */
+export type AzureVmWorkloadSAPHanaDBInstance = AzureVmWorkloadProtectableItem & {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  protectableItemType: "SAPHanaDBInstance";
+};
+
+/** Azure VM workload-specific protectable item representing SAP HANA Dbinstance. */
+export type AzureVmWorkloadSAPHanaHSR = AzureVmWorkloadProtectableItem & {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  protectableItemType: "SAPHanaHSR";
 };
 
 /** Azure VM workload-specific protectable item representing SQL Availability Group. */
@@ -3619,7 +3721,8 @@ export enum KnownDataSourceType {
   SQLDataBase = "SQLDataBase",
   AzureFileShare = "AzureFileShare",
   SAPHanaDatabase = "SAPHanaDatabase",
-  SAPAseDatabase = "SAPAseDatabase"
+  SAPAseDatabase = "SAPAseDatabase",
+  SAPHanaDBInstance = "SAPHanaDBInstance"
 }
 
 /**
@@ -3641,7 +3744,8 @@ export enum KnownDataSourceType {
  * **SQLDataBase** \
  * **AzureFileShare** \
  * **SAPHanaDatabase** \
- * **SAPAseDatabase**
+ * **SAPAseDatabase** \
+ * **SAPHanaDBInstance**
  */
 export type DataSourceType = string;
 
@@ -4017,54 +4121,6 @@ export enum KnownBackupEngineType {
  */
 export type BackupEngineType = string;
 
-/** Known values of {@link ContainerType} that the service accepts. */
-export enum KnownContainerType {
-  Invalid = "Invalid",
-  Unknown = "Unknown",
-  IaasVMContainer = "IaasVMContainer",
-  IaasVMServiceContainer = "IaasVMServiceContainer",
-  DPMContainer = "DPMContainer",
-  AzureBackupServerContainer = "AzureBackupServerContainer",
-  MABContainer = "MABContainer",
-  Cluster = "Cluster",
-  AzureSqlContainer = "AzureSqlContainer",
-  Windows = "Windows",
-  VCenter = "VCenter",
-  VMAppContainer = "VMAppContainer",
-  SqlagWorkLoadContainer = "SQLAGWorkLoadContainer",
-  StorageContainer = "StorageContainer",
-  GenericContainer = "GenericContainer",
-  MicrosoftClassicComputeVirtualMachines = "Microsoft.ClassicCompute/virtualMachines",
-  MicrosoftComputeVirtualMachines = "Microsoft.Compute/virtualMachines",
-  AzureWorkloadContainer = "AzureWorkloadContainer"
-}
-
-/**
- * Defines values for ContainerType. \
- * {@link KnownContainerType} can be used interchangeably with ContainerType,
- *  this enum contains the known values that the service supports.
- * ### Known values supported by the service
- * **Invalid** \
- * **Unknown** \
- * **IaasVMContainer** \
- * **IaasVMServiceContainer** \
- * **DPMContainer** \
- * **AzureBackupServerContainer** \
- * **MABContainer** \
- * **Cluster** \
- * **AzureSqlContainer** \
- * **Windows** \
- * **VCenter** \
- * **VMAppContainer** \
- * **SQLAGWorkLoadContainer** \
- * **StorageContainer** \
- * **GenericContainer** \
- * **Microsoft.ClassicCompute\/virtualMachines** \
- * **Microsoft.Compute\/virtualMachines** \
- * **AzureWorkloadContainer**
- */
-export type ContainerType = string;
-
 /** Known values of {@link ProtectionState} that the service accepts. */
 export enum KnownProtectionState {
   Invalid = "Invalid",
@@ -4191,7 +4247,8 @@ export enum KnownWorkloadType {
   SQLDataBase = "SQLDataBase",
   AzureFileShare = "AzureFileShare",
   SAPHanaDatabase = "SAPHanaDatabase",
-  SAPAseDatabase = "SAPAseDatabase"
+  SAPAseDatabase = "SAPAseDatabase",
+  SAPHanaDBInstance = "SAPHanaDBInstance"
 }
 
 /**
@@ -4213,7 +4270,8 @@ export enum KnownWorkloadType {
  * **SQLDataBase** \
  * **AzureFileShare** \
  * **SAPHanaDatabase** \
- * **SAPAseDatabase**
+ * **SAPAseDatabase** \
+ * **SAPHanaDBInstance**
  */
 export type WorkloadType = string;
 
@@ -4224,7 +4282,9 @@ export enum KnownPolicyType {
   Differential = "Differential",
   Log = "Log",
   CopyOnlyFull = "CopyOnlyFull",
-  Incremental = "Incremental"
+  Incremental = "Incremental",
+  SnapshotFull = "SnapshotFull",
+  SnapshotCopyOnlyFull = "SnapshotCopyOnlyFull"
 }
 
 /**
@@ -4237,9 +4297,53 @@ export enum KnownPolicyType {
  * **Differential** \
  * **Log** \
  * **CopyOnlyFull** \
- * **Incremental**
+ * **Incremental** \
+ * **SnapshotFull** \
+ * **SnapshotCopyOnlyFull**
  */
 export type PolicyType = string;
+
+/** Known values of {@link TieringMode} that the service accepts. */
+export enum KnownTieringMode {
+  Invalid = "Invalid",
+  TierRecommended = "TierRecommended",
+  TierAfter = "TierAfter",
+  DoNotTier = "DoNotTier"
+}
+
+/**
+ * Defines values for TieringMode. \
+ * {@link KnownTieringMode} can be used interchangeably with TieringMode,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Invalid** \
+ * **TierRecommended** \
+ * **TierAfter** \
+ * **DoNotTier**
+ */
+export type TieringMode = string;
+
+/** Known values of {@link RetentionDurationType} that the service accepts. */
+export enum KnownRetentionDurationType {
+  Invalid = "Invalid",
+  Days = "Days",
+  Weeks = "Weeks",
+  Months = "Months",
+  Years = "Years"
+}
+
+/**
+ * Defines values for RetentionDurationType. \
+ * {@link KnownRetentionDurationType} can be used interchangeably with RetentionDurationType,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Invalid** \
+ * **Days** \
+ * **Weeks** \
+ * **Months** \
+ * **Years**
+ */
+export type RetentionDurationType = string;
 
 /** Known values of {@link HealthStatus} that the service accepts. */
 export enum KnownHealthStatus {
@@ -4351,7 +4455,9 @@ export enum KnownRestorePointType {
   Full = "Full",
   Log = "Log",
   Differential = "Differential",
-  Incremental = "Incremental"
+  Incremental = "Incremental",
+  SnapshotFull = "SnapshotFull",
+  SnapshotCopyOnlyFull = "SnapshotCopyOnlyFull"
 }
 
 /**
@@ -4363,7 +4469,9 @@ export enum KnownRestorePointType {
  * **Full** \
  * **Log** \
  * **Differential** \
- * **Incremental**
+ * **Incremental** \
+ * **SnapshotFull** \
+ * **SnapshotCopyOnlyFull**
  */
 export type RestorePointType = string;
 
@@ -4429,7 +4537,9 @@ export enum KnownRestorePointQueryType {
   Differential = "Differential",
   FullAndDifferential = "FullAndDifferential",
   All = "All",
-  Incremental = "Incremental"
+  Incremental = "Incremental",
+  SnapshotFull = "SnapshotFull",
+  SnapshotCopyOnlyFull = "SnapshotCopyOnlyFull"
 }
 
 /**
@@ -4443,31 +4553,11 @@ export enum KnownRestorePointQueryType {
  * **Differential** \
  * **FullAndDifferential** \
  * **All** \
- * **Incremental**
+ * **Incremental** \
+ * **SnapshotFull** \
+ * **SnapshotCopyOnlyFull**
  */
 export type RestorePointQueryType = string;
-
-/** Known values of {@link RetentionDurationType} that the service accepts. */
-export enum KnownRetentionDurationType {
-  Invalid = "Invalid",
-  Days = "Days",
-  Weeks = "Weeks",
-  Months = "Months",
-  Years = "Years"
-}
-
-/**
- * Defines values for RetentionDurationType. \
- * {@link KnownRetentionDurationType} can be used interchangeably with RetentionDurationType,
- *  this enum contains the known values that the service supports.
- * ### Known values supported by the service
- * **Invalid** \
- * **Days** \
- * **Weeks** \
- * **Months** \
- * **Years**
- */
-export type RetentionDurationType = string;
 
 /** Known values of {@link JobStatus} that the service accepts. */
 export enum KnownJobStatus {
@@ -4706,7 +4796,9 @@ export enum KnownBackupType {
   Differential = "Differential",
   Log = "Log",
   CopyOnlyFull = "CopyOnlyFull",
-  Incremental = "Incremental"
+  Incremental = "Incremental",
+  SnapshotFull = "SnapshotFull",
+  SnapshotCopyOnlyFull = "SnapshotCopyOnlyFull"
 }
 
 /**
@@ -4719,7 +4811,9 @@ export enum KnownBackupType {
  * **Differential** \
  * **Log** \
  * **CopyOnlyFull** \
- * **Incremental**
+ * **Incremental** \
+ * **SnapshotFull** \
+ * **SnapshotCopyOnlyFull**
  */
 export type BackupType = string;
 
@@ -4731,7 +4825,8 @@ export enum KnownWorkloadItemType {
   SAPHanaSystem = "SAPHanaSystem",
   SAPHanaDatabase = "SAPHanaDatabase",
   SAPAseSystem = "SAPAseSystem",
-  SAPAseDatabase = "SAPAseDatabase"
+  SAPAseDatabase = "SAPAseDatabase",
+  SAPHanaDBInstance = "SAPHanaDBInstance"
 }
 
 /**
@@ -4745,7 +4840,8 @@ export enum KnownWorkloadItemType {
  * **SAPHanaSystem** \
  * **SAPHanaDatabase** \
  * **SAPAseSystem** \
- * **SAPAseDatabase**
+ * **SAPAseDatabase** \
+ * **SAPHanaDBInstance**
  */
 export type WorkloadItemType = string;
 
@@ -4785,6 +4881,50 @@ export enum KnownIntentItemType {
  */
 export type IntentItemType = string;
 
+/** Known values of {@link ContainerType} that the service accepts. */
+export enum KnownContainerType {
+  Invalid = "Invalid",
+  Unknown = "Unknown",
+  IaasVMContainer = "IaasVMContainer",
+  IaasVMServiceContainer = "IaasVMServiceContainer",
+  DPMContainer = "DPMContainer",
+  AzureBackupServerContainer = "AzureBackupServerContainer",
+  MABContainer = "MABContainer",
+  Cluster = "Cluster",
+  AzureSqlContainer = "AzureSqlContainer",
+  Windows = "Windows",
+  VCenter = "VCenter",
+  VMAppContainer = "VMAppContainer",
+  SqlagWorkLoadContainer = "SQLAGWorkLoadContainer",
+  StorageContainer = "StorageContainer",
+  GenericContainer = "GenericContainer",
+  HanaHSRContainer = "HanaHSRContainer"
+}
+
+/**
+ * Defines values for ContainerType. \
+ * {@link KnownContainerType} can be used interchangeably with ContainerType,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Invalid** \
+ * **Unknown** \
+ * **IaasVMContainer** \
+ * **IaasVMServiceContainer** \
+ * **DPMContainer** \
+ * **AzureBackupServerContainer** \
+ * **MABContainer** \
+ * **Cluster** \
+ * **AzureSqlContainer** \
+ * **Windows** \
+ * **VCenter** \
+ * **VMAppContainer** \
+ * **SQLAGWorkLoadContainer** \
+ * **StorageContainer** \
+ * **GenericContainer** \
+ * **HanaHSRContainer**
+ */
+export type ContainerType = string;
+
 /** Known values of {@link BackupItemType} that the service accepts. */
 export enum KnownBackupItemType {
   Invalid = "Invalid",
@@ -4801,7 +4941,8 @@ export enum KnownBackupItemType {
   SQLDataBase = "SQLDataBase",
   AzureFileShare = "AzureFileShare",
   SAPHanaDatabase = "SAPHanaDatabase",
-  SAPAseDatabase = "SAPAseDatabase"
+  SAPAseDatabase = "SAPAseDatabase",
+  SAPHanaDBInstance = "SAPHanaDBInstance"
 }
 
 /**
@@ -4823,7 +4964,8 @@ export enum KnownBackupItemType {
  * **SQLDataBase** \
  * **AzureFileShare** \
  * **SAPHanaDatabase** \
- * **SAPAseDatabase**
+ * **SAPAseDatabase** \
+ * **SAPHanaDBInstance**
  */
 export type BackupItemType = string;
 
@@ -4891,6 +5033,26 @@ export type HttpStatusCode =
   | "ServiceUnavailable"
   | "GatewayTimeout"
   | "HttpVersionNotSupported";
+/** Defines values for ProtectableContainerType. */
+export type ProtectableContainerType =
+  | "Invalid"
+  | "Unknown"
+  | "IaasVMContainer"
+  | "IaasVMServiceContainer"
+  | "DPMContainer"
+  | "AzureBackupServerContainer"
+  | "MABContainer"
+  | "Cluster"
+  | "AzureSqlContainer"
+  | "Windows"
+  | "VCenter"
+  | "VMAppContainer"
+  | "SQLAGWorkLoadContainer"
+  | "StorageContainer"
+  | "GenericContainer"
+  | "Microsoft.ClassicCompute/virtualMachines"
+  | "Microsoft.Compute/virtualMachines"
+  | "AzureWorkloadContainer";
 /** Defines values for RecoveryPointTierType. */
 export type RecoveryPointTierType =
   | "Invalid"